feat(src): :rocket: Implementing heredoc process and various utils in the project
Here, a new folder "utils" was created. It is a garbage folder, as it will be greatly modified on merging branches to main (as my partner and I need the same kind of stuff). Also, the heredoc_process is implemented, working (it should), and need to be connected to the io_env_manager.